requirements.txt
boto3
flask
mysql-connector-python
werkzeug